Sépare les réductions pour les investissements d'outremer des autres réductions
